requirements.txt
pyserial>=3.5
construct>=2.10.68
pyyaml>=6.0
